re-arranged domain hirearchy to be more domain centered
